+++ changelog.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.0...HEAD)
+### Fixed
+* [PCM-1836](https://inindca.atlassian.net/browse/PCM-1836) - Remove circular ref issue when logging the session when updating default devices.
+
 ### Changed
 * [PCM-1387](https://inindca.atlassian.net/browse/PCM-1387) - Made screenRecordingMetadata a required param for acceptSession for screen recording sessions; Those metadatas get sent to the server automatically.