+++ changelog.md
@@ -4,13 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.0...HEAD)
+
+# [v7.1.1](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.1.0...v7.1.1)
 ### Fixed
 * [PCM-1836](https://inindca.atlassian.net/browse/PCM-1836) - Remove circular ref issue when logging the session when updating default devices.
 
 ### Changed
 * [PCM-1387](https://inindca.atlassian.net/browse/PCM-1387) - Made screenRecordingMetadata a required param for acceptSession for screen recording sessions; Those metadatas get sent to the server automatically.
-
-### Added
 * [PCM-1279](https://inindca.atlassian.net/browse/PCM-1279) – calling `sdk.setAccessToken()` will now also pass
     that token down to the streaming-client and the client-logger.